+++ CHANGELOG.md
@@ -12,3 +12,5 @@ throw `PabloK\SupercacheBundle\Exceptions\EmptyPathException`
 when you pass empty `$cacheDir`
 - Removed `beberlei/assert` dependency
 - Added `PabloK\SupercacheBundle\Cache\CacheType` class for type constants
+- Added `PabloK\SupercacheBundle\Factory\ResponseFactory` class for creating new cache responses
+from `PabloK\SupercacheBundle\Cache\CacheElement`